Refactor: Update setOcrModel method to accept additional data types
The setOcrModel method in the TextDetector class has been modified to accept not only OcrModelIndex and str data types, but also int. This change allows for more flexibility when specifying the OCR model to be used. Additionally, the method now logs the selected OCR model and the corresponding model folder.